+++ packages/expo-video/CHANGELOG.md
@@ -7,6 +7,8 @@
 ### 🎉 New features
 
 - [Android][iOS] Added an option to `generateThumbnailsAsync` to set the maximum size of generated thumbnails. ([#33599](https://github.com/expo/expo/pull/33599) by [@tsapeta](https://github.com/tsapeta)), ([#33712](https://github.com/expo/expo/pull/33712) by [@behenate](https://github.com/behenate))
+- [Android][iOS] Add support for listing available video tracks and currently played video track.
+- [Android][iOS] Add `sourceLoad` event.
 
 ### 🐛 Bug fixes